+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. `BooleanExp
 
 ---
 
+## [1.0.1](https://github.com/ABridoux/BooleanExpressionEvaluation/tree/1.0.0) (28/11/2019)
+
+### Added
+
+- Conformance to codable for `ExpressionElement` and thus for `Expression` too
+- `Expression` description and evaluation
+
 ## [1.0.0](https://github.com/ABridoux/BooleanExpressionEvaluation/tree/1.0.0) (27/11/2019)
 
 Initiial release